Final Website to turn in
I added in the contact info to the footer. Removed the code I put in to try and make it responsive but that made it worse. I'm not sure what happened to the social media icons in the footer. They were originally round, I recopied the code as it originally was and they were still oval.